Extract user menu to its own component
